+++ src/Adapters/BlinkerMQTTAT.h
@@ -69,7 +69,7 @@ class BlinkerMQTTAT
         void begin(const char* auth);
         bool autoPrint(uint32_t id);
         // bool autoPrint(char *name, char *type, char *data);
-        // bool autoPrint(char *name1, char *type1, char *data1, \
+        // bool autoPrint(char *name1, char *type1, char *data1, 
         //             char *name2, char *type2, char *data2);
         char * deviceName();
         char * authKey() { return _authKey; }
@@ -84,7 +84,7 @@ class BlinkerMQTTAT
         bool init() { return isMQTTinit; }
         bool reRegister() { return connectServer(); }
         void aligenieType(blinker_at_aligenie_t _type) { _aliType = _type; }
-        void aligenieType(blinker_at_dueros_t _type) { _duerType = _type; }
+        void duerType(blinker_at_dueros_t _type) { _duerType = _type; }
         void freshAlive() { kaTime = millis(); isAlive = true; }
 
     private :
@@ -94,10 +94,12 @@ class BlinkerMQTTAT
         void mDNSInit();
         void checkKA();
         bool checkAliKA();
+        bool checkDuerKA();
         bool checkCanPrint();
         bool checkCanBprint();
         bool checkPrintSpan();
         bool checkAliPrintSpan();
+        bool checkDuerPrintSpan();
 
     protected :
         blinker_at_aligenie_t   _aliType = ALI_NONE;
@@ -129,10 +131,15 @@ class BlinkerMQTTAT
         uint32_t    respTime = 0;
         uint8_t     respAliTimes = 0;
         uint32_t    respAliTime = 0;
+        uint8_t     respDuerTimes = 0;
+        uint32_t    respDuerTime = 0;
 
         uint32_t    aliKaTime = 0;
         bool        isAliAlive = false;
         bool        isAliAvail = false;
+        uint32_t    duerKaTime = 0;
+        bool        isDuerAlive = false;
+        bool        isDuerAvail = false;
         char*       mqtt_broker;
 
         Stream*     stream;
@@ -249,6 +256,8 @@ void webSocketEvent_MQTT_AT(uint8_t num, WStype_t type, \
 
             // send message to client
             // webSocket_MQTT_AT.sendBIN(num, payload, length);
+            break;
+        default :
             break;
     }
 }
@@ -928,17 +937,17 @@ bool BlinkerMQTTAT::bPrint(char * name, const String & data)
 
         String bPubTopic = BLINKER_F("");
 
-        if (mqtt_broker == BLINKER_MQTT_BORKER_ONENET)
-        {
-            bPubTopic = MQTT_PRODUCTINFO_MQTT_AT;
-            bPubTopic += BLINKER_F("/");
-            bPubTopic += name;
-            bPubTopic += BLINKER_F("/r");
-        }
-        else
-        {
-            bPubTopic = BLINKER_PUB_TOPIC_MQTT_AT;
-        }
+        // if (mqtt_broker == BLINKER_MQTT_BORKER_ONENET)
+        // {
+        //     bPubTopic = MQTT_PRODUCTINFO_MQTT_AT;
+        //     bPubTopic += BLINKER_F("/");
+        //     bPubTopic += name;
+        //     bPubTopic += BLINKER_F("/r");
+        // }
+        // else
+        // {
+        //     bPubTopic = BLINKER_PUB_TOPIC_MQTT_AT;
+        // }
 
         if (! mqtt_MQTT_AT->publish(bPubTopic.c_str(), data_add.c_str()))
         {
@@ -1033,7 +1042,7 @@ bool BlinkerMQTTAT::duerPrint(const String & data)
 
     data_add += data;
     data_add += BLINKER_F(",\"fromDevice\":\"");
-    data_add += MQTT_ID_MQTT;
+    data_add += MQTT_ID_MQTT_AT;
     data_add += BLINKER_F("\",\"toDevice\":\"DuerOS_r\"");
     data_add += BLINKER_F(",\"deviceType\":\"vAssistant\"}");
 
@@ -1042,7 +1051,7 @@ bool BlinkerMQTTAT::duerPrint(const String & data)
     BLINKER_LOG_ALL(BLINKER_F("MQTT DuerOS Publish..."));
     BLINKER_LOG_FreeHeap_ALL();
 
-    if (mqtt_MQTT->connected())
+    if (mqtt_MQTT_AT->connected())
     {
         if (!checkDuerKA())
         {
@@ -1056,7 +1065,7 @@ bool BlinkerMQTTAT::duerPrint(const String & data)
         }
         respDuerTime = millis();
 
-        if (! mqtt_MQTT->publish(BLINKER_PUB_TOPIC_MQTT, data_add.c_str()))
+        if (! mqtt_MQTT_AT->publish(BLINKER_PUB_TOPIC_MQTT_AT, data_add.c_str()))
         {
             BLINKER_LOG_ALL(data_add);
             BLINKER_LOG_ALL(BLINKER_F("...Failed"));
@@ -1205,7 +1214,7 @@ bool BlinkerMQTTAT::autoPrint(uint32_t id)
 //     BLINKER_LOG_ALL(BLINKER_F("autoPrint..."));
     
 //     if (mqtt_MQTT_AT->connected()) {
-//         if ((millis() - linkTime) > BLINKER_LINK_MSG_LIMIT || \
+//         if ((millis() - linkTime) > BLINKER_LINK_MSG_LIMIT || 
 //             linkTime == 0)
 //         {
 //             linkTime = millis();
@@ -1244,7 +1253,7 @@ bool BlinkerMQTTAT::autoPrint(uint32_t id)
     
 //     if (mqtt_MQTT_AT->connected())
 //     {
-//         if ((millis() - linkTime) > BLINKER_LINK_MSG_LIMIT || \
+//         if ((millis() - linkTime) > BLINKER_LINK_MSG_LIMIT || 
 //             linkTime == 0)
 //         {
 //             linkTime = millis();
@@ -1675,33 +1684,33 @@ bool BlinkerMQTTAT::connectServer() {
 
 #elif defined(ESP32)
     String host = BLINKER_F("https://iotdev.clz.me");
-    // const char* ca = \ 
-    //     "-----BEGIN CERTIFICATE-----\n" \
-    //     "MIIEgDCCA2igAwIBAgIQDKTfhr9lmWbWUT0hjX36oDANBgkqhkiG9w0BAQsFADBy\n" \
-    //     "MQswCQYDVQQGEwJDTjElMCMGA1UEChMcVHJ1c3RBc2lhIFRlY2hub2xvZ2llcywg\n" \
-    //     "SW5jLjEdMBsGA1UECxMURG9tYWluIFZhbGlkYXRlZCBTU0wxHTAbBgNVBAMTFFRy\n" \
-    //     "dXN0QXNpYSBUTFMgUlNBIENBMB4XDTE4MDEwNDAwMDAwMFoXDTE5MDEwNDEyMDAw\n" \
-    //     "MFowGDEWMBQGA1UEAxMNaW90ZGV2LmNsei5tZTCCASIwDQYJKoZIhvcNAQEBBQAD\n" \
-    //     "ggEPADCCAQoCggEBALbOFn7cJ2I/FKMJqIaEr38n4kCuJCCeNf1bWdWvOizmU2A8\n" \
-    //     "QeTAr5e6Q3GKeJRdPnc8xXhqkTm4LOhgdZB8KzuVZARtu23D4vj4sVzxgC/zwJlZ\n" \
-    //     "MRMxN+cqI37kXE8gGKW46l2H9vcukylJX+cx/tjWDfS2YuyXdFuS/RjhCxLgXzbS\n" \
-    //     "cve1W0oBZnBPRSMV0kgxTWj7hEGZNWKIzK95BSCiMN59b+XEu3NWGRb/VzSAiJEy\n" \
-    //     "Hy9DcDPBC9TEg+p5itHtdMhy2gq1OwsPgl9HUT0xmDATSNEV2RB3vwviNfu9/Eif\n" \
-    //     "ObhsV078zf30TqdiESqISEB68gJ0Otru67ePoTkCAwEAAaOCAWowggFmMB8GA1Ud\n" \
-    //     "IwQYMBaAFH/TmfOgRw4xAFZWIo63zJ7dygGKMB0GA1UdDgQWBBR/KLqnke61779P\n" \
-    //     "xc9htonQwLOxPDAYBgNVHREEETAPgg1pb3RkZXYuY2x6Lm1lMA4GA1UdDwEB/wQE\n" \
-    //     "AwIFoDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwTAYDVR0gBEUwQzA3\n" \
-    //     "BglghkgBhv1sAQIwKjAoBggrBgEFBQcCARYcaHR0cHM6Ly93d3cuZGlnaWNlcnQu\n" \
-    //     "Y29tL0NQUzAIBgZngQwBAgEwgYEGCCsGAQUFBwEBBHUwczAlBggrBgEFBQcwAYYZ\n" \
-    //     "aHR0cDovL29jc3AyLmRpZ2ljZXJ0LmNvbTBKBggrBgEFBQcwAoY+aHR0cDovL2Nh\n" \
-    //     "Y2VydHMuZGlnaXRhbGNlcnR2YWxpZGF0aW9uLmNvbS9UcnVzdEFzaWFUTFNSU0FD\n" \
-    //     "QS5jcnQwCQYDVR0TBAIwADANBgkqhkiG9w0BAQsFAAOCAQEAhtM4eyrWB14ajJpQ\n" \
-    //     "ibZ5FbzVuvv2Le0FOSoss7UFCDJUYiz2LiV8yOhL4KTY+oVVkqHaYtcFS1CYZNzj\n" \
-    //     "6xWcqYZJ+pgsto3WBEgNEEe0uLSiTW6M10hm0LFW9Det3k8fqwSlljqMha3gkpZ6\n" \
-    //     "8WB0f2clXOuC+f1SxAOymnGUsSqbU0eFSgevcOIBKR7Hr3YXBXH3jjED76Q52OMS\n" \
-    //     "ucfOM9/HB3jN8o/ioQbkI7xyd/DUQtzK6hSArEoYRl3p5H2P4fr9XqmpoZV3i3gQ\n" \
-    //     "oOdVycVtpLunyUoVAB2DcOElfDxxXCvDH3XsgoIU216VY03MCaUZf7kZ2GiNL+UX\n" \
-    //     "9UBd0Q==\n" \
+    // const char* ca = 
+    //     "-----BEGIN CERTIFICATE-----\n" 
+    //     "MIIEgDCCA2igAwIBAgIQDKTfhr9lmWbWUT0hjX36oDANBgkqhkiG9w0BAQsFADBy\n" 
+    //     "MQswCQYDVQQGEwJDTjElMCMGA1UEChMcVHJ1c3RBc2lhIFRlY2hub2xvZ2llcywg\n" 
+    //     "SW5jLjEdMBsGA1UECxMURG9tYWluIFZhbGlkYXRlZCBTU0wxHTAbBgNVBAMTFFRy\n" 
+    //     "dXN0QXNpYSBUTFMgUlNBIENBMB4XDTE4MDEwNDAwMDAwMFoXDTE5MDEwNDEyMDAw\n" 
+    //     "MFowGDEWMBQGA1UEAxMNaW90ZGV2LmNsei5tZTCCASIwDQYJKoZIhvcNAQEBBQAD\n" 
+    //     "ggEPADCCAQoCggEBALbOFn7cJ2I/FKMJqIaEr38n4kCuJCCeNf1bWdWvOizmU2A8\n" 
+    //     "QeTAr5e6Q3GKeJRdPnc8xXhqkTm4LOhgdZB8KzuVZARtu23D4vj4sVzxgC/zwJlZ\n" 
+    //     "MRMxN+cqI37kXE8gGKW46l2H9vcukylJX+cx/tjWDfS2YuyXdFuS/RjhCxLgXzbS\n" 
+    //     "cve1W0oBZnBPRSMV0kgxTWj7hEGZNWKIzK95BSCiMN59b+XEu3NWGRb/VzSAiJEy\n" 
+    //     "Hy9DcDPBC9TEg+p5itHtdMhy2gq1OwsPgl9HUT0xmDATSNEV2RB3vwviNfu9/Eif\n" 
+    //     "ObhsV078zf30TqdiESqISEB68gJ0Otru67ePoTkCAwEAAaOCAWowggFmMB8GA1Ud\n" 
+    //     "IwQYMBaAFH/TmfOgRw4xAFZWIo63zJ7dygGKMB0GA1UdDgQWBBR/KLqnke61779P\n" 
+    //     "xc9htonQwLOxPDAYBgNVHREEETAPgg1pb3RkZXYuY2x6Lm1lMA4GA1UdDwEB/wQE\n" 
+    //     "AwIFoDAdBgNVHSUEFjAUBggrBgEFBQcDAQYIKwYBBQUHAwIwTAYDVR0gBEUwQzA3\n" 
+    //     "BglghkgBhv1sAQIwKjAoBggrBgEFBQcCARYcaHR0cHM6Ly93d3cuZGlnaWNlcnQu\n" 
+    //     "Y29tL0NQUzAIBgZngQwBAgEwgYEGCCsGAQUFBwEBBHUwczAlBggrBgEFBQcwAYYZ\n" 
+    //     "aHR0cDovL29jc3AyLmRpZ2ljZXJ0LmNvbTBKBggrBgEFBQcwAoY+aHR0cDovL2Nh\n" 
+    //     "Y2VydHMuZGlnaXRhbGNlcnR2YWxpZGF0aW9uLmNvbS9UcnVzdEFzaWFUTFNSU0FD\n" 
+    //     "QS5jcnQwCQYDVR0TBAIwADANBgkqhkiG9w0BAQsFAAOCAQEAhtM4eyrWB14ajJpQ\n" 
+    //     "ibZ5FbzVuvv2Le0FOSoss7UFCDJUYiz2LiV8yOhL4KTY+oVVkqHaYtcFS1CYZNzj\n" 
+    //     "6xWcqYZJ+pgsto3WBEgNEEe0uLSiTW6M10hm0LFW9Det3k8fqwSlljqMha3gkpZ6\n" 
+    //     "8WB0f2clXOuC+f1SxAOymnGUsSqbU0eFSgevcOIBKR7Hr3YXBXH3jjED76Q52OMS\n" 
+    //     "ucfOM9/HB3jN8o/ioQbkI7xyd/DUQtzK6hSArEoYRl3p5H2P4fr9XqmpoZV3i3gQ\n" 
+    //     "oOdVycVtpLunyUoVAB2DcOElfDxxXCvDH3XsgoIU216VY03MCaUZf7kZ2GiNL+UX\n" 
+    //     "9UBd0Q==\n" 
     //     "-----END CERTIFICATE-----\n";
 // #endif
 
@@ -2042,6 +2051,13 @@ bool BlinkerMQTTAT::checkAliKA() {
         return true;
 }
 
+bool BlinkerMQTTAT::checkDuerKA() {
+    if (millis() - duerKaTime >= 10000)
+        return false;
+    else
+        return true;
+}
+
 bool BlinkerMQTTAT::checkCanPrint() {
     if ((millis() - printTime >= BLINKER_MQTT_MSG_LIMIT && isAlive) || printTime == 0) {
         return true;
@@ -2103,6 +2119,29 @@ bool BlinkerMQTTAT::checkAliPrintSpan()
     else
     {
         respAliTimes = 0;
+        return true;
+    }
+}
+
+bool BlinkerMQTTAT::checkDuerPrintSpan()
+{
+    if (millis() - respDuerTime < BLINKER_PRINT_MSG_LIMIT/2)
+    {
+        if (respDuerTimes > BLINKER_PRINT_MSG_LIMIT/2)
+        {
+            BLINKER_ERR_LOG(BLINKER_F("DUEROS NOT ALIVE OR MSG LIMIT"));
+            
+            return false;
+        }
+        else
+        {
+            respDuerTimes++;
+            return true;
+        }
+    }
+    else
+    {
+        respDuerTimes = 0;
         return true;
     }
 }
